SCC-2158 Initial Commit of (generic) holding poller
This is the initial commit of a function for polling for record updates from the Sierra API. It is intended to be configurable to handle any bib, item or holdings records but is intended to be used with holdings data first.

It fetches batches of updates and stores the current state of its progress in S3, in the event of an error, the function will resume at its previous position in the update stream.

Once fetched records are encoded with an Avro schema and passed to a Kinesis stream.

If over 1% of records are unable to be processed an error is raised (along with other standard erros).

Immediate development priorities are:

- Unit test suite
- Better documentation
- Integration testing with all components