Fix race condition processing final set of results
Fix race condition that may drop records when kinesis write time exceeds Sierra fetch time:

When more than 100 results are received, the poller must fetch a second set of records. Writing the first set of records to kinesis occurs in parallel to fetching that second set of records. If Sierra takes N time to return the second set of records, but it takes N + 1 time to write the first set of records to Kinesis, the app (previously) nullified the second result set (held in @previous_results), making it impossible for the app to subsequently send the second result set through kinesis. Essentially, the app assumes that fetching records from Sierra will always be slower than writing 100 records to Kinesis, which is not always true.

The fix is to recast our @previous_records as an array, so that it can grow to hold as many as two buffered resultsets if Kinesis is slow to write. In the scenario above, instead of nullifying @previous_results after writing the records to Kinesis, the kinesis-writing code shifts the resultset off of @previous_results right before beginning the kinesis-write, allowing the Sierra-retreiving code to safely push additional resultsets onto @previous_results without risk that they'll be erroneously nullified.